extract Swift closure wrapping;
